+++ CHANGES.rst
@@ -58,6 +58,8 @@ Features added
   Patch by Adam Turner.
 * #13647: LaTeX: allow more cases of table nesting.
   Patch by Jean-François B.
+* #13657: LaTeX: support CSS3 length units.
+  Patch by Jean-François B.
 * #13684: intersphinx: Add a file-based cache for remote inventories.
   The location of the cache directory must not be relied upon externally,
   as it may change without notice or warning in future releases.